@@ -2,7 +2,10 @@
   "name": "hire-tabs",
   "version": "2.0.2",
   "description": "",
-	"keywords": ["react-component", "tabs"],
+  "keywords": [
+    "react-component",
+    "tabs"
+  ],
   "main": "build/index.js",
   "scripts": {
     "start": "scripts/watch.sh",
@@ -18,24 +21,13 @@
     "url": "https://github.com/Hire-Forms/hire-tabs/issues"
   },
   "homepage": "https://github.com/Hire-Forms/hire-tabs",
-  "peerDependencies": {
-    "classnames": "^2.1.2",
-    "react": "^0.13.3"
-  },
   "devDependencies": {
-    "babel-core": "^5.5.8",
-    "babel-eslint": "^3.1.15",
-    "babel-plugin-object-assign": "^1.1.0",
-    "babelify": "^6.1.2",
-    "browserify": "^10.2.4",
-    "derequire": "^2.0.0",
-    "eslint": "^0.23.0",
-    "eslint-plugin-react": "^2.5.2",
+    "@types/react": "^15.0.15",
     "hi-bump": "^1.1.1",
-    "watchify": "^3.3.1"
+    "typescript": "^2.2.1"
   },
   "dependencies": {
-    "hire-forms-prop-types": "^1.0.1",
-    "hire-forms-utils": "^1.0.1"
+    "classnames": "^2.2.5",
+    "react": "^15.4.2"
   }
 }
